backend/requirements.txt
# Core
fastapi>=0.100.0,<1.0.0
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.28.0

# Redis
redis==5.0.1
aioredis==2.0.1
ratelimit==2.2.1

# Monitoring and Metrics
sentry-sdk[fastapi]==1.35.0
prometheus-client==0.18.0
prometheus-fastapi-instrumentator>=6.0.0,<7.0.0
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-exporter-otlp==1.20.0
opentelemetry-instrumentation-fastapi==0.41b0
opentelemetry-instrumentation-sqlalchemy==0.41b0
opentelemetry-instrumentation-redis==0.41b0
opentelemetry-instrumentation-logging==0.41b0
opentelemetry-semantic-conventions==0.41b0

# Utils
python-multipart==0.0.6
email-validator==2.1.0.post1
python-slugify==8.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1

# CORS
starlette>=0.27.0,<1.0.0

# Email
aiosmtplib==2.0.1
email-validator==2.1.0.post1
